requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
faster-whisper==1.0.3
transformers==4.45.0
torch==2.5.0
piper-tts==1.2.0
numpy==1.26.4
soundfile==0.12.1
webrtcvad==2.0.10
